edgeservice/requirements.txt
requests==2.24.0
django==3.1.1
djangorestframework==3.11.1
django-cors-headers==3.5.0